+++ CHANGELOG.md
@@ -18,3 +18,5 @@ Changelog
 ### Features
 
 * Add favicons and touch icons.
+* Store and return "updatedAt" value with bookmarks. This string tracks the
+  date and time that the bookmark was last updated.